hfi core: queues setup using RPC queue APIs
queues setup using RPC queue APIs

+++ hfi_core/inc/hfi_queue.h
@@ -17,7 +17,7 @@
  */
 struct hfi_queue_buffer {
 	u64 kva;
-	u32 dva;
+	u64 dva;
 	u32 buf_len;
 };
